extract UserDropdownContent into a single component
