+++ site/src/components/UserDropdown/UsersDropdown.tsx
@@ -1,26 +1,14 @@
 import Badge from "@material-ui/core/Badge"
-import Divider from "@material-ui/core/Divider"
-import ListItemIcon from "@material-ui/core/ListItemIcon"
-import ListItemText from "@material-ui/core/ListItemText"
 import MenuItem from "@material-ui/core/MenuItem"
 import { fade, makeStyles } from "@material-ui/core/styles"
-import AccountIcon from "@material-ui/icons/AccountCircleOutlined"
 import React, { useState } from "react"
-import { Link } from "react-router-dom"
 import * as TypesGen from "../../api/typesGenerated"
 import { navHeight } from "../../theme/constants"
 import { BorderedMenu } from "../BorderedMenu/BorderedMenu"
 import { CloseDropdown, OpenDropdown } from "../DropdownArrows/DropdownArrows"
-import { DocsIcon } from "../Icons/DocsIcon"
-import { LogoutIcon } from "../Icons/LogoutIcon"
 import { UserAvatar } from "../UserAvatar/UserAvatar"
-import { UserProfileCard } from "../UserProfileCard/UserProfileCard"
+import { UserDropdownContent } from "../UserDropdownContent/UserDropdownContent"
 
-export const Language = {
-  accountLabel: "Account",
-  docsLabel: "Documentation",
-  signOutLabel: "Sign Out",
-}
 export interface UserDropdownProps {
   user: TypesGen.User
   /**
@@ -68,43 +56,7 @@ export const UserDropdown: React.FC<UserDropdownProps> = ({ isOpen = false, user
         variant="user-dropdown"
         onClose={onPopoverClose}
       >
-        <div className={styles.userInfo}>
-          <UserProfileCard user={user} />
-
-          <Divider />
-
-          <Link to="/settings/account" className={styles.link}>
-            <MenuItem className={styles.menuItem} onClick={onPopoverClose}>
-              <ListItemIcon className={styles.icon}>
-                <AccountIcon />
-              </ListItemIcon>
-              <ListItemText primary={Language.accountLabel} />
-            </MenuItem>
-          </Link>
-
-          <Divider />
-
-          <a
-            href={`https://github.com/coder/coder/tree/${process.env.CODER_VERSION}/docs`}
-            target="_blank"
-            rel="noreferrer"
-            className={styles.link}
-          >
-            <MenuItem className={styles.menuItem} onClick={onPopoverClose}>
-              <ListItemIcon className={styles.icon}>
-                <DocsIcon />
-              </ListItemIcon>
-              <ListItemText primary={Language.docsLabel} />
-            </MenuItem>
-          </a>
-
-          <MenuItem className={styles.menuItem} onClick={onSignOut}>
-            <ListItemIcon className={styles.icon}>
-              <LogoutIcon />
-            </ListItemIcon>
-            <ListItemText primary={Language.signOutLabel} />
-          </MenuItem>
-        </div>
+        <UserDropdownContent user={user} onPopoverClose={onPopoverClose} onSignOut={onSignOut} />
       </BorderedMenu>
     </>
   )
@@ -123,10 +75,6 @@ export const useStyles = makeStyles((theme) => ({
     maxWidth: 300,
   },
 
-  userInfo: {
-    marginBottom: theme.spacing(1),
-  },
-
   menuItem: {
     height: navHeight,
     padding: `${theme.spacing(1.5)}px ${theme.spacing(2.75)}px`,
@@ -135,14 +83,5 @@ export const useStyles = makeStyles((theme) => ({
       backgroundColor: fade(theme.palette.primary.light, 0.1),
       transition: "background-color 0.3s ease",
     },
-  },
-
-  link: {
-    textDecoration: "none",
-    color: "inherit",
-  },
-
-  icon: {
-    color: theme.palette.text.secondary,
   },
 }))
